Extract setupTransactionService to separate file
