+++ universal-login-relayer/test/relayer/services/TransactionService.js
@@ -1,11 +1,12 @@
 import {expect} from 'chai';
 import {utils} from 'ethers';
 import {ACTION_KEY, calculateMessageSignature} from '@universal-login/contracts';
-import {setupTransactionService, transferMessage, addKeyMessage, removeKeyMessage} from '../../fixtures/basicWalletService';
+import {transferMessage, addKeyMessage, removeKeyMessage} from '../../fixtures/basicWalletService';
+import setupTransactionService from '../../helpers/basicTransactionService';
 import defaultDeviceInfo from '../../config/defaults';
 import {getKnex} from '../../../lib/utils/knexUtils';
 
-describe('Relayer - TransactionService', async () => {
+describe.only('Relayer - TransactionService', async () => {
   let transactionService;
   let provider;
   let authorisationService;
